fix: manage redundant code
